requirements.txt
altgraph==0.17
asgiref==3.4.1
astroid==2.3.3
autopep8==1.5.1
beautifulsoup4==4.9.3
black==21.10b0
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
colorama==0.4.3
cycler==0.10.0
Django==3.2.10
future==0.18.2
idna==2.10
importlib-metadata==4.0.1
isort==4.3.21
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.0
pathspec==0.9.0
pefile==2019.4.18
Pillow==7.1.2
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.5.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.5
pytz==2019.3
pywin32-ctypes==0.2.0
regex==2021.11.2
requests==2.25.1
selenium==3.141.0
six==1.14.0
soupsieve==2.1
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.2
wget==3.2
wrapt==1.11.2
zipp==3.4.1
